Shouldnt use symlinks - they get flattened by pypi
